# Create a prefetch class that wraps around prioritized replay to allow for prefetching of samples