--- a/pyrela/scripts/ref_run_apex.sh
+++ b/pyrela/scripts/ref_run_apex.sh
@@ -1,4 +1,4 @@
-export job_name='exps/apex_ms_pacman_prefetcher'
+export job_name='exps/apex_pong'
 mkdir -p ${job_name}
 sbatch --job-name $job_name \
        --mem 500G \
@@ -14,7 +14,7 @@ sbatch --job-name $job_name \
        --wrap "
 #!/bin/bash
 python main.py \
-       --game ms_pacman \
+       --game pong \
        --save_dir ${job_name} \
        --num_thread 80 \
        --num_game_per_thread 20 \
@@ -22,5 +22,4 @@ python main.py \
        --eps 1.5e-7 \
        --algo apex \
        --actor_sync_freq 20 \
-       --use_prefetcher True
 "