--- a/rela/prefetcher.h
+++ b/rela/prefetcher.h
@@ -3,8 +3,8 @@
 #include <atomic>
 #include <memory>
 #include <thread>
-#include <vector>
 #include <torch/script.h>
+#include <vector>
 
 #include "rela/prioritized_replay.h"
 
@@ -13,16 +13,12 @@ namespace rela {
 template <class PrioritizedReplay, class DataType>
 class Prefetcher {
  public:
-  Prefetcher(std::shared_ptr<PrioritizedReplay> replayer,
-             int batchsize)
-        : batchsize_(batchsize)
-        , replayer_(std::move(replayer)) {
-    sampleThr_ = std::thread([this]() {
-        workerThread();
-    });
+  Prefetcher(std::shared_ptr<PrioritizedReplay> replayer, int batchsize)
+      : batchsize_(batchsize)
+      , replayer_(std::move(replayer)) {
+    sampleThr_ = std::thread([this]() { workerThread(); });
     done_ = false;
     bufferSize_ = 50;
-    
   }
 
   std::tuple<DataType, torch::Tensor> sample() {
@@ -36,52 +32,55 @@ class Prefetcher {
   }
 
   void updatePriority(const torch::Tensor& priority) {
-    assert(priority.dim() == 1);    
-    assert(std::abs((int)sampledIndices_.size() - (int) sampleBuffer_.size()) < 3);
+    assert(priority.dim() == 1);
+    assert(std::abs((int)sampledIndices_.size() - (int)sampleBuffer_.size()) <
+           3);
 
     m_indices.lock();
-    std::vector<int> currIndices = sampledIndices_.front();    
+    std::vector<int> currIndices = sampledIndices_.front();
     sampledIndices_.pop_front();
     m_indices.unlock();
     replayer_->updatePrefetcherPriority(priority, currIndices);
 
     cv_fill_buffer.notify_one();
   }
- 
- ~Prefetcher() {
+
+  ~Prefetcher() {
     signalDone();
     wait();
     sampleThr_.join();
     sampledIndices_.clear();
     sampleBuffer_.clear();
- }
-
+  }
 
  private:
   void workerThread() {
-    while(true) {
-        wait_to_fill_buffer();
-        if (done_) {
-            cv_done.notify_one();
-            return;
-        }
-        sample_batch();
+    while (true) {
+      wait_to_fill_buffer();
+      if (done_) {
+        cv_done.notify_one();
+        return;
+      }
+      sample_batch();
     }
   }
 
   void wait_to_fill_buffer() {
-    std::unique_lock<std::mutex> lg (m_buffer);
-    while ((size_t)sampleBuffer_.size() > bufferSize_ || (size_t)replayer_->size() < 2 * batchsize_) {
-        if (done_) break;
-        cv_fill_buffer.wait(lg);
+    std::unique_lock<std::mutex> lg(m_buffer);
+    while ((size_t)sampleBuffer_.size() > bufferSize_ ||
+           (size_t)replayer_->size() < 2 * batchsize_) {
+      if (done_)
+        break;
+      cv_fill_buffer.wait(lg);
     }
   }
 
   void wait_for_buffer_to_fill() {
-    if ((size_t)replayer_->size() >= 2 * batchsize_) cv_fill_buffer.notify_one();
-    std::unique_lock<std::mutex> lg (m_buffer);
+    if ((size_t)replayer_->size() >= 2 * batchsize_)
+      cv_fill_buffer.notify_one();
+    std::unique_lock<std::mutex> lg(m_buffer);
     while (sampleBuffer_.size() == 0) {
-        cv_buffer_empty.wait(lg);
+      cv_buffer_empty.wait(lg);
     }
   }
 
@@ -91,7 +90,7 @@ class Prefetcher {
     m_indices.lock();
     sampledIndices_.push_back(indices);
     m_indices.unlock();
-    
+
     m_buffer.lock();
     sampleBuffer_.push_back(batch);
     m_buffer.unlock();
@@ -99,7 +98,7 @@ class Prefetcher {
     replayer_->clearSampledIndices();
     cv_buffer_empty.notify_one();
   }
-  
+
   void signalDone() {
     done_ = true;
     cv_fill_buffer.notify_one();
@@ -107,14 +106,13 @@ class Prefetcher {
 
   void wait() {
     std::unique_lock<std::mutex> lg(m_buffer);
-    while(!done_) {
-        cv_done.wait(lg);
+    while (!done_) {
+      cv_done.wait(lg);
     }
   }
-  
+
   std::deque<std::vector<int>> sampledIndices_;
   std::deque<std::tuple<DataType, torch::Tensor>> sampleBuffer_;
-  
 
   size_t batchsize_;
   size_t bufferSize_;
@@ -130,7 +128,6 @@ class Prefetcher {
 
   std::shared_ptr<PrioritizedReplay> replayer_;
   std::thread sampleThr_;
-
 };
 
 using FFPrefetcher = Prefetcher<FFPrioritizedReplay, FFTransition>;