Old Sample Timings
@@@Time
	sync and updating: 2 MS, 4.61%
	sample data      : 21 MS, 40.96%
	device conversion: 6 MS, 12.78%
	calculating loss : 7 MS, 15.03%
	backprop & update: 13 MS, 26.46%
	updating priority: 0 MS, 0.16%
@@@total time per iter: 51.44 ms
epoch 25, eval score: 713.000000
[25] Time spent = 59.32 s

New Sample Timings
@@@Time
	sync and updating: 0 MS, 1.37%
	sample data      : 39 MS, 60.95%
	device conversion: 5 MS, 8.40%
	calculating loss : 7 MS, 11.45%
	backprop & update: 10 MS, 16.13%
	updating priority: 1 MS, 1.69%
@@@total time per iter: 65.27 ms
epoch 26, eval score: 800.000000
[26] Time spent = 76.31 s

For the asteroids game, we are able to decrease data sampling time by around 50%, and overall time spend per an epoch decreased is ~25%. Will update later with more complete statistics.